feat: update demo for new APIs
