--- TrinityAudioDemo.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,14 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "trinityaudio-ios-sdk",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/TrinityAudioSDK/trinityaudio-ios-sdk",
-      "state" : {
-        "branch" : "main",
-        "revision" : "6517b8781a14542ad762127cacbb3974f67b4931"
-      }
-    }
-  ],
-  "version" : 2
-}
